refactor pubkey, signature, and proof of possession modules
